backend/requirements.txt
# Framework web
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic==2.5.3

# OCR et traitement de documents
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
PyMuPDF>=1.23.0
